requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-openai
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain_community